Revert "Split Validation Rules"
This reverts commit fc7e18ceab2d543e566fb3e24e1fb12e690506bf.
