--- system/Validation/Rules/Length.php
@@ -1,106 +1,0 @@
-<?php namespace CodeIgniter\Validation\Rules;
-
-/**
- * CodeIgniter
- *
- * An open source application development framework for PHP
- *
- * This content is released under the MIT License (MIT)
- *
- * Copyright (c) 2014 - 2017, British Columbia Institute of Technology
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @package	CodeIgniter
- * @author	CodeIgniter Dev Team
- * @copyright	Copyright (c) 2014 - 2017, British Columbia Institute of Technology (http://bcit.ca/)
- * @license	https://opensource.org/licenses/MIT	MIT License
- * @link	https://codeigniter.com
- * @since	Version 3.0.0
- * @filesource
- */
-
-/**
- * Rules.
- *
- * @package CodeIgniter\Validation
- */
-class Length
-{
-	/**
-	 * Returns true if $str is $val characters long.
-	 *
-	 * @param string $str
-	 * @param string $val
-	 * @param array  $data
-	 *
-	 * @return bool
-	 */
-	public function exact_length(string $str=null, string $val, array $data): bool
-	{
-		if (! is_numeric($val))
-		{
-			return false;
-		}
-
-		return ((int)$val == mb_strlen($str));
-	}
-
-	//--------------------------------------------------------------------
-
-	/**
-	 * Returns true if $str is $val or fewer characters in length.
-	 *
-	 * @param string $str
-	 * @param string $val
-	 * @param array  $data
-	 *
-	 * @return bool
-	 */
-	public function max_length(string $str=null, string $val, array $data): bool
-	{
-		if (! is_numeric($val))
-		{
-			return false;
-		}
-
-		return ($val >= mb_strlen($str));
-	}
-
-	//--------------------------------------------------------------------
-
-	/**
-	 * Returns true if $str is at least $val length.
-	 *
-	 * @param string $str
-	 * @param string $val
-	 * @param array  $data
-	 *
-	 * @return bool
-	 */
-	public function min_length(string $str=null, string $val, array $data): bool
-	{
-		if (! is_numeric($val))
-		{
-			return false;
-		}
-
-		return ($val <= mb_strlen($str));
-	}
-}
