+++ tests/system/Validation/ValidationTest.php
@@ -10,29 +10,22 @@ class ValidationTest extends \CIUnitTestCase
 	 */
 	protected $validation;
 
-	protected $config = [
-		'ruleSets'      => [
-			\CodeIgniter\Validation\Rules\Required::class,
-			\CodeIgniter\Validation\Rules\AlphaFormat::class,
-			\CodeIgniter\Validation\Rules\DateTimeFormat::class,
-			\CodeIgniter\Validation\Rules\NumberFormat::class,
-			\CodeIgniter\Validation\Rules\Format::class,
-			\CodeIgniter\Validation\Rules\Comparison::class,
-			\CodeIgniter\Validation\Rules\DatabaseDependency::class,
-			\CodeIgniter\Validation\Rules\Length::class,
-			\CodeIgniter\Validation\Rules\File::class,
-			\CodeIgniter\Validation\Rules\CreditCard::class,
-			\CodeIgniter\Validation\TestRules::class,
-		],
-		'groupA'        => [
-			'foo' => 'required|min_length[5]',
-		],
-		'groupA_errors' => [
-			'foo' => [
-				'min_length' => 'Shame, shame. Too short.',
-			],
-		],
-	];
+    protected $config = [
+        'ruleSets'      => [
+            \CodeIgniter\Validation\Rules::class,
+            \CodeIgniter\Validation\FileRules::class,
+            \CodeIgniter\Validation\CreditCardRules::class,
+            \CodeIgniter\Validation\TestRules::class,
+        ],
+        'groupA'        => [
+            'foo' => 'required|min_length[5]',
+        ],
+        'groupA_errors' => [
+            'foo' => [
+                'min_length' => 'Shame, shame. Too short.',
+            ],
+        ],
+    ];
 
 	//--------------------------------------------------------------------
